added a custom alert dialog and listView click listener. when movie in list is clicked, brings up popup with more info (release date and rating)

+++ app/src/main/java/com/example/sarch/echefilm/utilities/Movie.java
@@ -11,7 +11,7 @@ import java.util.ArrayList;
 public class Movie implements Serializable {
 
     private int id;
-    private  int voteAverage;
+    private  float voteAverage;
     private int voteCount;
     private JSONArray genres;
     private String originalTitle;
@@ -38,7 +38,7 @@ public class Movie implements Serializable {
         this.id = id;
     }
 
-    public int getVoteAverage() {
+    public float getVoteAverage() {
         return voteAverage;
     }
